requirements.txt
tensorflow>=2.10.0
torch>=1.12.0
torchvision>=0.13.0
scikit-learn>=1.2.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
joblib>=1.2.0
Pillow>=9.3.0
opencv-python>=4.7.0
mediapipe>=0.10.0
tqdm>=4.65.0
pandas>=1.5.0
